Carregamento, confirm ao apagar e pag de error

+++ Semana 10 - Custom Hooks/labex/src/functions/deleteTrip.js
@@ -3,10 +3,15 @@ import { urlApi, headersConfig } from '../apiConfig/axiosConfig'
 
 
 const deleteTrip = (id) => {
-    axios.delete(`${urlApi}/trips/${id}`, headersConfig)
-    .then((res) => { 
+    const res = window.confirm('Tem certeza que deseja apagar a viagem?')
+    if (res) {
+        axios.delete(`${urlApi}/trips/${id}`, headersConfig)
+            .then((res) => {
+            }).catch((err) => {
+                alert('Viagem não apagada')
+            })
+    }
 
-    }).catch((err) => { })
 }
 
 export default deleteTrip
